+++ src/Chirp.Razor/CheepRepo/ICheepRepo.cs
@@ -1,0 +1,7 @@
+namespace Chirp.SQLite.CheepRepos;
+public interface ICheepRepo 
+{
+    public List<CheepViewModel> GetCheeps(int page);
+
+    public List<CheepViewModel> GetCheepsFromAuthor(string author, int page);
+}
